Methods refactored to use the models changes.
Methods Word.objects.(get, search...) now use a lexicon code to do it,
or use the variables src_language and dst_language to do those methods.

Changed all references to Lexicon.name to Lexicon.code

importvariation.py file now uses a lexicon code to perform the Word.objects. methods
and also need a lexicon code as a parameter when using it as a command